ai_counseling/requirements.txt
google-generativeai==0.3.1
supabase==1.0.3
python-dotenv==0.19.2
Flask==2.0.1
Flask-Login==0.5.0

